requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
discord.py==2.5.2
exceptiongroup==1.3.0
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
multidict==6.4.3
ollama==0.4.8
propcache==0.3.1
pydantic==2.11.4
pydantic_core==2.33.2
sniffio==1.3.1
typing-inspection==0.4.0
typing_extensions==4.13.2
yarl==1.20.0
